% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{from_lookup}
\alias{from_lookup}
\title{Fetch a lookup table and return it.}
\usage{
from_lookup(keys, value, version = NULL, source = NULL)
}
\arguments{
\item{keys}{A single lookup table key or a vector of keys.}

\item{value}{A lookup table value.}

\item{version}{A lookup table version identifier.}

\item{source}{A \code{\link{lookup_source}} or \code{NULL} to use the default.}
}
\value{
A \code{tbl}, or \code{NULL} if the table does not exist.
}
\description{
Fetch a lookup table and return it.
}
