% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_metadata}
\alias{get_metadata}
\title{Fetch remote metadata for a lookup table.}
\usage{
get_metadata(x, keys, value, version = NULL)
}
\arguments{
\item{x}{A \code{\link{lookup_source}} object.}

\item{keys}{A single lookup table key or a vector of keys.}

\item{value}{A lookup table value.}

\item{version}{A lookup table version identifier.}
}
\value{
See \code{\link[yaml]{yaml.load}}.
}
\description{
Fetch remote metadata for a lookup table.
}
