% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_table}
\alias{get_table}
\title{Fetch a lookup table from a remote url.}
\usage{
get_table(x, keys, value, version = NULL)
}
\arguments{
\item{x}{A \code{\link{lookup_source}} object.}

\item{keys}{A single lookup table key or a vector of keys.}

\item{value}{A lookup table value.}

\item{version}{A lookup table version identifier.}
}
\value{
A \code{tbl}, or \code{NULL} if the table does not exist.
}
\description{
Fetch a lookup table from a remote url.
}
