% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lookup}
\alias{lookup}
\title{Fetch a lookup table and join it to the give \code{data.frame}.}
\usage{
lookup(data, keys, value, lookup_keys = NULL, version = NULL,
  source = NULL, require_match = NULL)
}
\arguments{
\item{data}{A \code{data.frame} to join to.}

\item{keys}{A key column or vector of key columns to join on.}

\item{value}{A lookup table value.}

\item{lookup_keys}{A single lookup table key or a vector of keys, if different from \code{keys}.}

\item{version}{A lookup table version identifier.}

\item{source}{A \code{lookup_source} or \code{NULL} to use the default.}

\item{require_match}{If \code{TRUE}, values without a match in the lookup table will raise an error.}
}
\value{
A \code{tbl}, or \code{NULL} if the table does not exist.
}
