% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lookup_cpi}
\alias{lookup_cpi}
\title{Fetch the CPI lookup table and use it to inflation-adjust columns in a dataset.}
\usage{
lookup_cpi(data, values, year = "year", base = "2016",
  replace_columns = TRUE, source = NULL)
}
\arguments{
\item{data}{A \code{data.frame} to join to.}

\item{values}{A single column name to be adjusteed or a vector of column names.}

\item{year}{The name of a column containing data years.}

\item{base}{The base year to adjust data to.}

\item{replace_columns}{If \code{TRUE}, the original, unadjusted columns will be replaced.}

\item{source}{A \code{\link{lookup_source}} or \code{NULL} to use the default.}
}
\value{
A \code{tbl}, or \code{NULL} if the table does not exist.
}
\description{
This function works with \href{https://CRAN.R-project.org/package=magrittr}{magrittr} pipes.
}
