% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{lookup_source}
\alias{lookup_source}
\title{Create a \code{lookup_source} object which handles fetching lookup tables.}
\usage{
lookup_source(root = "http://wireservice.github.io/lookup")
}
\arguments{
\item{root}{Root url to the hosted lookup tables.}
}
\value{
A source object.
}
\description{
Create a \code{lookup_source} object which handles fetching lookup tables.
}
