% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupr.R
\docType{package}
\name{lookupr}
\alias{lookupr}
\alias{lookupr-package}
\title{lookupr}
\description{
Lookup pulls common lookup tables from a Github repository
and joins them to your data.
}
\details{
This is a port of Python's \href{https://agate-lookup.readthedocs.io}{agate-lookup}.

This version does not do caching. If you don't want to redownload
lookup tables, save a local copy of your results.
}
