% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{make_metadata_path}
\alias{make_metadata_path}
\title{Generate a path to find a given lookup table's metadata.}
\usage{
make_metadata_path(keys, value, version = NULL)
}
\arguments{
\item{keys}{A single lookup table key or a vector of keys.}

\item{value}{A lookup table value.}

\item{version}{A lookup table version identifier.}
}
\value{
A lookup table metadata path.
}
\description{
Generate a path to find a given lookup table's metadata.
}
