% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{make_table_path}
\alias{make_table_path}
\title{Generate a path to find a given lookup table.}
\usage{
make_table_path(keys, value, version = NULL)
}
\arguments{
\item{keys}{A single lookup table key or a vector of keys.}

\item{value}{A lookup table value.}

\item{version}{A lookup table version identifier.}
}
\value{
A lookup table path.
}
\description{
Generate a path to find a given lookup table.
}
