% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{map_column_type}
\alias{map_column_type}
\title{Convert agate column types to R types.}
\usage{
map_column_type(c)
}
\arguments{
\item{c}{A column definition from lookup table metadata.}
}
\value{
A column type suitable for use with \code{\link[readr]{read_csv}}.
}
\description{
Convert agate column types to R types.
}
