% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lookup}
\alias{lookup}
\title{Fetch a lookup table and join it to your data.}
\usage{
lookup(data, keys, value, lookup_keys = NULL, version = NULL,
  source = NULL)
}
\arguments{
\item{data}{A \code{data.frame} to join to.}

\item{keys}{A key column or vector of key columns to join on.}

\item{value}{A lookup table value.}

\item{lookup_keys}{A single lookup table key or a vector of keys, if different from \code{keys}.}

\item{version}{A lookup table version identifier.}

\item{source}{A \code{\link{lookup_source}} or \code{NULL} to use the default.}
}
\value{
A \code{tbl}, or \code{NULL} if the table does not exist.
}
\description{
This function works with \href{https://CRAN.R-project.org/package=magrittr}{magrittr} pipes.
}
